Vibrator: update sequence of triggering composePlayThread
Currently, a bool flag inComposition is used to identify if the
composePlayThread is in progress or not, and it is set in the
parent thread after starting the child thread and unset after
the child thread is run completely. There is a chance that the
child thread would run completely prior to the parent thread setting
the flag, in that case the flag will be still true even after the
child thread has already done the composition and result in a while(true)
loop.

To fix this, set the inComposition before starting the child thread,
and also update the while(inComposition) loop with a timeout value
according to the play length of the composition effect.

Change-Id: If7591e0776eb2593c4187990f99536f6916e63b0